doc: File Handler (aka metahandler and childs)
